worker/requirements.txt
redis
minio
psycopg2-binary
sqlalchemy
python-dotenv
ultralytics
opencv-python-headless
Pillow
